+++ src/colors.rs
@@ -1,0 +1,9 @@
+pub const COLORS: &[colored::Color] = &[
+    colored::Color::Green,
+    colored::Color::Yellow,
+    colored::Color::Blue,
+    colored::Color::Magenta,
+    colored::Color::Cyan,
+    colored::Color::White,
+    colored::Color::BrightMagenta,
+];
