+++ src/main.rs
@@ -1,27 +1,22 @@
-use crate::parser::load_config;
+use crate::{colors::COLORS, parser::load_config};
 
+mod colors;
 mod config;
+mod kill_process;
 mod parser;
 mod print_banner;
 mod runner;
+mod spawn_service;
+mod watcher;
 
 #[tokio::main]
 async fn main() {
     clearscreen::clear().expect("Failed to clear the screen");
     let config = load_config("fyrer.yml");
-    let colors = vec![
-        colored::Color::Green,
-        colored::Color::Yellow,
-        colored::Color::Blue,
-        colored::Color::Magenta,
-        colored::Color::Cyan,
-        colored::Color::White,
-        colored::Color::BrightMagenta,
-    ];
     let mut handles = vec![];
     print_banner::print_banner();
     for (i, service) in config.services.into_iter().enumerate() {
-        let color = colors[i % colors.len()];
+        let color = COLORS[i % COLORS.len()];
         let handle = tokio::spawn(runner::runner(service, color));
         handles.push(handle);
     }
